ChatBotPlatform/requirements.txt
# Core FastAPI and async
fastapi==0.115.0
uvicorn[standard]==0.37.0
pydantic==2.10.0
pydantic-settings==2.6.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17

# HTTP client for external APIs
httpx==0.28.0

# Data processing and RAG
numpy==1.26.4
pandas==2.2.3
faiss-cpu==1.9.0
openai==1.54.0
tiktoken==0.8.0
sentence-transformers==3.2.1
trafilatura==1.12.2

# File processing
PyPDF2==3.0.1
python-docx==1.1.2
python-magic==0.4.27

# Utilities
python-dotenv==1.0.1
cryptography==43.0.3

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.28.0  # For testing FastAPI

# Development
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.13.0

# Logging and monitoring
structlog==24.4.0
rich==13.9.4

# Rate limiting
slowapi==0.1.9

# CORS
python-cors==1.7.0